backend/requirements.txt
_libgcc_mutex==0.1
_openmp_mutex==5.1
absl-py==2.1.0
anaconda-anon-usage==0.4.4
anyio==4.6.2.post1
archspec==0.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
boltons==23.0.0
brotli-python==1.0.9
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2024.3.11
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
comm==0.2.2
conda==24.4.0
conda-content-trust==0.2.0
conda-libmamba-solver==24.1.0
conda-package-handling==2.2.0
conda-package-streaming==0.9.0
contourpy==1.3.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.1.0
expat==2.6.2
fastjsonschema==2.20.0
filelock==3.16.1
fmt==9.1.0
fonttools==4.55.0
fqdn==1.5.1
fsspec==2024.10.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
icu==73.1
idna==3.7
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.4
json5==0.9.28
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.1
jupyterlab-language-pack-zh-cn==4.2.post3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
krb5==1.20.1
ld_impl_linux-64==2.38
libarchive==3.6.2
libcurl==8.7.1
libedit==3.1.20230828
libev==4.33
libffi==3.4.4
libgcc-ng==11.2.0
libgomp==11.2.0
libmamba==1.5.8
libmambapy==1.5.8
libnghttp2==1.57.0
libsolv==0.7.24
libssh2==1.11.0
libstdcxx-ng==11.2.0
libuuid==1.41.5
libxml2==2.10.4
lz4-c==1.9.4
markdown==3.7
markupsafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
menuinst==2.0.2
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ncurses==6.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook-shim==0.2.4
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
openssl==3.0.13
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pcre2==10.42
pexpect==4.9.0
pillow==11.0.0
pip==24.0
platformdirs==3.10.0
pluggy==1.0.0
prometheus-client==0.21.0
prompt-toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pybind11-abi==5
pycosat==0.6.6
pycparser==2.21
pygments==2.18.0
pyparsing==3.2.0
pysocks==1.7.1
python==3.12.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyyaml==6.0.2
pyzmq==26.2.0
readline==8.2
referencing==0.35.1
reproc==14.2.4
reproc-cpp==14.2.4
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
ruamel.yaml==0.17.21
send2trash==1.8.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlite==3.45.3
stack-data==0.6.3
supervisor==4.2.5
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
tinycss2==1.4.0
tk==8.6.14
torch==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.2
tqdm==4.66.2
traitlets==5.14.3
triton==3.1.0
truststore==0.8.0
types-python-dateutil==2.9.0.20241003
typing-extensions==4.12.2
tzdata==2024a
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.13
xz==5.4.6
yaml-cpp==0.8.0
zlib==1.2.13
zstandard==0.22.0
zstd==1.5.5
flask==2.0.1
flask-cors==3.0.10
psutil==5.9.0
pynvml==11.5.0
opencv-python==4.5.3.56
numpy==1.21.2
torch==1.9.0
torchvision==0.10.0
